rename・fix : なぜなら、「コード規約に従う」ために、「ディレクトリ名を変更し、インポートを修正」した
