+++ src/front/src/views/components/blocks/header/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Button } from '../../atoms/Button/Button';
+import { Button } from '../../atoms/button/Button';
 import './header.css';
 
 type User = {
