+++ src/front/src/views/pages/page/Page.stories.ts
